backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Admin Panel
sqladmin==0.16.1
itsdangerous==2.1.2

# Pydantic and validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography>=41.0.0  # For Fernet encryption

# Environment variables
python-dotenv==1.0.0

# WebSockets
websockets==12.0

# Redis (for caching and pub/sub)
redis==5.0.1
hiredis==2.2.3

# AI and ML
httpx==0.25.2
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
PyPDF2==3.0.1

# Additional utilities
python-dateutil==2.8.2
